ref: move shared toml.ts lib
